feat!: Rework schema compatibility
This new implementation is (in my opinion) simpler than the previous
iteration and easier to modify when needed.

I've introduced the concept of partial and full compatibility. Partial
compatibility means that there incompatiblities but that depends on whats
written, for example an enum `[A, B, C]` is partially compatible with
the enum `[A, B]` but depends on what the writer has written.

Users wishing to be absolutely sure, can treat `Compatibility::Partial`
as an error, others can use it to slowly upgrade from an old schema.